+++ goldenverba/components/chunking/CodeChunker.py
@@ -21,7 +21,7 @@ class CodeChunker(Chunker):
     def __init__(self):
         super().__init__()
         self.name = "Code"
-        self.requires_library = ["langchain_text_splitters "]
+        self.requires_library = ["langchain_text_splitters"]
         self.description = "Split code based on programming language using LangChain"
         self.config = {
             "Language": InputConfig(
@@ -56,17 +56,19 @@ class CodeChunker(Chunker):
         chunk_size = config["Chunk Size"].value
         chunk_overlap = config["Chunk Overlap"].value
 
-        text_splitter = RecursiveCharacterTextSplitter.from_language(language=language, chunk_size=chunk_size, chunk_overlap=chunk_overlap)
+        text_splitter = RecursiveCharacterTextSplitter.from_language(
+            language=language, chunk_size=chunk_size, chunk_overlap=chunk_overlap
+        )
 
         for document in documents:
 
             # Skip if document already contains chunks
             if len(document.chunks) > 0:
                 continue
-            
+
             char_end_i = -1
             for i, chunk in enumerate(text_splitter.split_text(document.content)):
-                
+
                 if chunk_overlap == 0:
                     char_start_i = char_end_i + 1
                     char_end_i = char_start_i + len(chunk)