+++ goldenverba/components/generation/GroqGenerator.py
@@ -179,6 +179,8 @@ def get_models(url: str, api_key: str) -> List[str]:
         ]
         models.sort()
         models = filter_models(models)
+        if len(models) == 0:
+            return DEFAULT_MODEL_LIST
         return models
 
     except Exception as e: